fix: allow setting the tenant when creating new periodic task
This makes it possible to add a new PeriodicTask while on a Tenant-level
admin and set the Tenant on the PeriodicTaskTenantLink correctly.
Previously, you couldn't set the Tenant because it is readonly, which then
meant that once you saved the PeriodicTask, it was on the public schema,
and therefore wouldn't show up on the Tenant's admin.